Refactor Notifications object
